requirements.txt
streamlit
pyautogen
openai
pandas
matplotlib
gspread
oauth2client